* commit 'b46ccc28852e0fabba23767b075f4a14db5a8698':
  ART: Fix a static_cast int32_t -> uint64_t bug.

+++ compiler/optimizing/nodes.h
@@ -2239,7 +2239,9 @@ class HIntConstant : public HConstant {
  public:
   int32_t GetValue() const { return value_; }
 
-  uint64_t GetValueAsUint64() const OVERRIDE { return static_cast<uint64_t>(value_); }
+  uint64_t GetValueAsUint64() const OVERRIDE {
+    return static_cast<uint64_t>(static_cast<uint32_t>(value_));
+  }
 
   bool InstructionDataEquals(HInstruction* other) const OVERRIDE {
     DCHECK(other->IsIntConstant());
